// src/app/peerpharm/schedule/models/schedule.ts
export class Schedule {
  scheduleId: string;
  qty: string;
  batch: string;
  costumer: string;
  date: Date;
  item: string;
  marks: string;
  mkp: string;
  orderN: string;
  packageP: string;
  positionN: number;
  productName: string;
  qtyProduced: string;
  shift: string;
  status: string;
  productionLine: string;
  pLinePositionN: number;
  cmptsStatus: string;
}